services/log-parser/requirements.txt
kafka-python==2.0.2
typing-extensions==4.8.0